Move alerce/antares data generation
